Refactor ControlledGates into CircPool and Decomposition
